refactor(editor): extract useEditorViewport (Phase 1c)
Move viewport geometry — viewportRef, measured viewportSize (ResizeObserver),
base placement, image rect, and client→viewport coord mapping — into
useEditorViewport({ open, transformedPreview, editorState }). transformedPreview
stays in EditorOverlay (it bridges image + transform and feeds save/frame too),
and the initial-crop-snapshot effect stays (it records into history).

Completes Phase 1 (core: image / history / viewport). Behavior-preserving:
21 editor/golden e2e green (03/06/07 + 18/19/20).
